Win32 File I/O; size_t stream position/size; improvements
Stream and its subclasses now keep track of the stream position
and size using `size_t` and not `long` -- negative values are
not needed and `size_t` is 64-bit compatible.

Implemented a Win32-specific File class that accesses files
directly using the Win32 file API. This should have better
performance compared to MSYS's POSIX file layer, but of course
it will behave like Windows files. Text mode is handled manually
by replacing `\r\n` pairs with `\n`. Stream position reflects
the actual bytes in the file rather than the text content.

In the standard C version there was a bug keeping track of the
stream position when it came to `\r\n` replacements. (Use the
test in t_test.c that reads single bytes from a file.)

+++ include/the_Foundation/buffer.h
@@ -66,15 +66,15 @@ iLocalDef iStream *     stream_Buffer   (iBuffer *d) { return &d->stream; }
 
 void        clear_Buffer    (iBuffer *);
 
-iLocalDef iBool isOpen_Buffer (const iBuffer *d) { return d->data != NULL; }
-iLocalDef int   mode_Buffer   (const iBuffer *d) { return d->mode ;}
-iLocalDef long  pos_Buffer    (const iBuffer *d) { return pos_Stream(&d->stream); }
-iLocalDef long  size_Buffer   (const iBuffer *d) { return size_Stream(&d->stream); }
-iLocalDef iBool isEmpty_Buffer(const iBuffer *d) { return size_Buffer(d) == 0; }
-iLocalDef iBool atEnd_Buffer  (const iBuffer *d) { return atEnd_Stream(&(d)->stream); }
+iLocalDef iBool  isOpen_Buffer (const iBuffer *d) { return d->data != NULL; }
+iLocalDef int    mode_Buffer   (const iBuffer *d) { return d->mode ;}
+iLocalDef size_t pos_Buffer    (const iBuffer *d) { return pos_Stream(&d->stream); }
+iLocalDef size_t size_Buffer   (const iBuffer *d) { return size_Stream(&d->stream); }
+iLocalDef iBool  isEmpty_Buffer(const iBuffer *d) { return size_Buffer(d) == 0; }
+iLocalDef iBool  atEnd_Buffer  (const iBuffer *d) { return atEnd_Stream(&(d)->stream); }
 
 iLocalDef void          rewind_Buffer       (iBuffer *d) { seek_Stream(&d->stream, 0); }
-iLocalDef void          seek_Buffer         (iBuffer *d, long offset) { seek_Stream(&d->stream, offset); }
+iLocalDef void          seek_Buffer         (iBuffer *d, size_t offset) { seek_Stream(&d->stream, offset); }
 iLocalDef void          write32_Buffer      (iBuffer *d, int32_t value) { write32_Stream(&d->stream, value); }
 iLocalDef size_t        writeData_Buffer    (iBuffer *d, const void *data, size_t size) { return writeData_Stream(&d->stream, data, size); }
